Time: 161 ms (59.77%), Space: 75.5 MB (60.34%) - LeetHub

+++ 1953-maximum-number-of-weeks-for-which-you-can-work/1953-maximum-number-of-weeks-for-which-you-can-work.cpp
@@ -7,7 +7,7 @@ public:
             sum+=i;mx=max(mx,i);
         }
         long long rest=sum-mx;
-        if(mx>rest){return ((long long)2*rest)+1;}
+        if(mx>rest){return 2*rest+1;}
         return sum;
     }
 };
